Rename cosmjs-provider file
